# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API do Sistema de Análises de Crédito (autenticação, listagem e SDUI).",
        "title": "Credit Analysis API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Autentica o usuário e retorna um token JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Autenticação",
                "parameters": [
                    {
                        "description": "Credenciais",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.loginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/credit-analyses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista análises de crédito com filtros, paginação e ordenação.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credit-analyses"
                ],
                "summary": "Listar análises",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtra por CPF/CNPJ",
                        "name": "document",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtra por nome do cliente",
                        "name": "clientName",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "APROVADO",
                            "REPROVADO",
                            "EM_ANALISE",
                            "PENDENTE"
                        ],
                        "type": "string",
                        "description": "Filtra por status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Score mínimo",
                        "name": "scoreMin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Score máximo",
                        "name": "scoreMax",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Data inicial (RFC3339)",
                        "name": "dateFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Data final (RFC3339)",
                        "name": "dateTo",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Página (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Itens por página (default 20)",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Campo de ordenação (default createdAt)",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Direção",
                        "name": "sortDir",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Page-model_CreditAnalysis"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria uma nova análise de crédito e emula o processo de análise (PENDENTE → EM_ANALISE → APROVADO/REPROVADO), retornando o resultado final com o histórico de eventos.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credit-analyses"
                ],
                "summary": "Criar análise",
                "parameters": [
                    {
                        "description": "Dados do cliente",
                        "name": "analysis",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createAnalysisRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.CreditAnalysisDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/credit-analyses/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna uma análise de crédito com seu histórico de eventos.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credit-analyses"
                ],
                "summary": "Detalhe da análise",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da análise",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CreditAnalysisDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Verifica se o serviço está no ar.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/sdui/screens/credit-analyses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna o contrato SDUI que descreve a tela de listagem de análises.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sdui"
                ],
                "summary": "SDUI da tela de listagem",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sdui.Screen"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/sdui/screens/login": {
            "get": {
                "description": "Retorna o contrato SDUI que descreve a tela de login.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sdui"
                ],
                "summary": "SDUI da tela de login",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sdui.Screen"
                        }
                    }
                }
            }
        },
        "/users/preferences/filters": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Salva as preferências de filtros do usuário autenticado (JSON arbitrário).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "preferences"
                ],
                "summary": "Salvar preferências de filtros",
                "parameters": [
                    {
                        "description": "Estado dos filtros",
                        "name": "filters",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.createAnalysisRequest": {
            "type": "object",
            "required": [
                "clientName",
                "document"
            ],
            "properties": {
                "clientName": {
                    "type": "string",
                    "minLength": 2
                },
                "document": {
                    "type": "string",
                    "minLength": 11
                }
            }
        },
        "handler.loginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 4
                }
            }
        },
        "handler.loginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "model.AnalysisStatus": {
            "type": "string",
            "enum": [
                "APROVADO",
                "REPROVADO",
                "EM_ANALISE",
                "PENDENTE"
            ],
            "x-enum-varnames": [
                "StatusApproved",
                "StatusRejected",
                "StatusInReview",
                "StatusPending"
            ]
        },
        "model.CreditAnalysis": {
            "type": "object",
            "properties": {
                "clientName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "document": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.AnalysisStatus"
                }
            }
        },
        "model.CreditAnalysisDetail": {
            "type": "object",
            "properties": {
                "clientName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "document": {
                    "type": "string"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CreditAnalysisEvent"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.AnalysisStatus"
                }
            }
        },
        "model.CreditAnalysisEvent": {
            "type": "object",
            "properties": {
                "analysisId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.AnalysisStatus"
                }
            }
        },
        "model.Page-model_CreditAnalysis": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CreditAnalysis"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "sdui.Column": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                }
            }
        },
        "sdui.Component": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sdui.Column"
                    }
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sdui.Field"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "sdui.Field": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "sdui.Screen": {
            "type": "object",
            "properties": {
                "components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sdui.Component"
                    }
                },
                "screen": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Informe o token como: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}