# backend/internal/handler/auth.go
package handler

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

type loginRequest struct {
	Email    string `json:"email" validate:"required,email"`
	Password string `json:"password" validate:"required,min=4"`
}

type loginResponse struct {
	Token string `json:"token"`
}

// Login authenticates the user and returns a JWT.
// @Summary      Autenticação
// @Description  Autentica o usuário e retorna um token JWT.
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        credentials  body      loginRequest   true  "Credenciais"
// @Success      200          {object}  loginResponse
// @Failure      400          {object}  map[string]string
// @Failure      401          {object}  map[string]string
// @Router       /auth/login [post]
func (h *Handler) Login(c *gin.Context) {
	var req loginRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid payload"})
		return
	}
	if err := h.validate.Struct(req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	token, err := h.auth.Login(c.Request.Context(), req.Email, req.Password)
	if err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "invalid credentials"})
		return
	}
	c.JSON(http.StatusOK, loginResponse{Token: token})
}